+++ Scroller/EditableWrapper.js
@@ -4,11 +4,17 @@ import {is} from '@enact/core/keymap';
 import {mergeClassNameMaps} from '@enact/core/util';
 import Spotlight from '@enact/spotlight';
 import Accelerator from '@enact/spotlight/Accelerator';
+import {Announce} from '@enact/ui/AnnounceDecorator';
 import classNames from 'classnames';
+import IString from 'ilib/lib/IString';
 import PropTypes from 'prop-types';
 import {useCallback, useEffect, useRef} from 'react';
 
+import $L from '../internal/$L';
+
 import componentCss from './EditableWrapper.module.less';
+
+const completeAnnounceDelay = 300; // An arbitrary delay at a level that is not ignored by the new focus element
 
 /**
  * The shape for editable of [Scroller]{@link sandstone/Scroller}.
@@ -65,6 +71,7 @@ const EditableWrapper = (props) => {
 
 		// DOM elements
 		selectedItem: null,
+		selectedItemLabel: '',
 		rearrangedItems: [],
 
 		// Indices
@@ -81,8 +88,11 @@ const EditableWrapper = (props) => {
 		lastMouseClientX: null,
 
 		// Last InputType which moves Items
-		lastInputType: null
+		lastInputType: null,
+		lastInputDirection: null
 	});
+
+	const announceRef = useRef({});
 
 	// Functions
 
@@ -93,6 +103,7 @@ const EditableWrapper = (props) => {
 		selectedItem?.classList.remove(componentCss.selected, customCss.selected, componentCss.rearranged);
 
 		mutableRef.current.selectedItem = null;
+		mutableRef.current.selectedItemLabel = '';
 		mutableRef.current.lastMoveDirection = null;
 		mutableRef.current.prevToIndex = null;
 		wrapperRef.current.style.setProperty('--selected-item-offset', '0px');
@@ -145,9 +156,14 @@ const EditableWrapper = (props) => {
 
 			item.classList.add(componentCss.selected, customCss.selected);
 			mutableRef.current.selectedItem = item;
+			mutableRef.current.selectedItemLabel = (item.ariaLabel || item.textContent) + ' ';
 
 			mutableRef.current.fromIndex = Number(item.style.order) - 1;
 			mutableRef.current.prevToIndex = mutableRef.current.fromIndex;
+
+			announceRef.current.announce(
+				mutableRef.current.selectedItemLabel + $L('Move left and right or press up key to delete')
+			);
 		}
 	}, [customCss.selected]);
 
@@ -177,6 +193,23 @@ const EditableWrapper = (props) => {
 		ev.stopPropagation();
 	}, [editable, finalizeOrders, findItemNode, reset, startEditing]);
 
+	const readOutCurrentPosition = useCallback((neighborItem) => {
+		const {lastInputDirection, lastInputType, selectedItemLabel} = mutableRef.current;
+		if (lastInputType === 'key') {
+			if (lastInputDirection === 'left') {
+				announceRef.current.announce(
+					new IString($L('{selectedItem} moved to the left of {neighborItem}')).format({selectedItem: selectedItemLabel, neighborItem}),
+					true
+				);
+			} else {
+				announceRef.current.announce(
+					new IString($L('{selectedItem} moved to the right of {neighborItem}')).format({selectedItem: selectedItemLabel, neighborItem}),
+					true
+				);
+			}
+		}
+	}, []);
+
 	// Add rearranged items
 	const addRearrangedItems = useCallback(({moveDirection, toIndex}) => {
 		// Set the moveDirection to css variable
@@ -186,6 +219,7 @@ const EditableWrapper = (props) => {
 		const {fromIndex, rearrangedItems, selectedItem} = mutableRef.current;
 		const getNextElement = (item) => moveDirection > 0 ? item.nextElementSibling : item.previousElementSibling;
 		let sibling = getNextElement(selectedItem);
+		let lastRearrangedItem;
 		let start = moveDirection > 0 ? toIndex : fromIndex;
 		let end = moveDirection > 0 ? fromIndex : toIndex;
 
@@ -196,23 +230,33 @@ const EditableWrapper = (props) => {
 				rearrangedItems.push(sibling);
 			}
 
+			lastRearrangedItem = sibling;
 			sibling = getNextElement(sibling);
 			start--;
 		}
 
+		if (lastRearrangedItem) {
+			readOutCurrentPosition(lastRearrangedItem.ariaLabel || lastRearrangedItem.textContent);
+		}
+
 		mutableRef.current.lastMoveDirection = moveDirection;
 
-	}, [scrollContainerHandle]);
+	}, [readOutCurrentPosition, scrollContainerHandle]);
 
 	const removeRearrangedItems = useCallback((numToRemove) => {
 		const {rearrangedItems} = mutableRef.current;
+		let toItem = null;
 		if (rearrangedItems.length > 0) {
 			for (let i = 0; i < numToRemove; i++) {
-				const toItem = rearrangedItems.pop();
+				toItem = rearrangedItems.pop();
 				toItem?.classList.remove(componentCss.rearrangedTransform);
 			}
 		}
-	}, []);
+
+		if (toItem) {
+			readOutCurrentPosition(toItem.ariaLabel || toItem.textContent);
+		}
+	}, [readOutCurrentPosition]);
 
 	// Move items
 	const moveItems = useCallback((toIndex) => {
@@ -276,8 +320,21 @@ const EditableWrapper = (props) => {
 		}
 
 		mutableRef.current.lastInputType = 'key';
+		mutableRef.current.lastInputDirection = is('left', keyCode) ? 'left' : 'right';
 		moveItems(toIndex);
-	}, [moveItems, scrollContainerHandle, scrollContentRef]);
+
+		if (toIndex <= 0) {
+			announceRef.current.announce(
+				(is('left', keyCode) && !rtl && $L('LEFTMOST')) ||
+				(is('right', keyCode) && rtl && $L('RIGHTMOST'))
+			);
+		} else if (toIndex >= dataSize - 1) {
+			announceRef.current.announce(
+				(is('right', keyCode) && !rtl &&  $L('RIGHTMOST')) ||
+				(is('left', keyCode) && rtl && $L('LEFTMOST'))
+			);
+		}
+	}, [dataSize, moveItems, scrollContainerHandle, scrollContentRef]);
 
 	// Remove an item
 	const removeItem = useCallback(() => {
@@ -342,7 +399,7 @@ const EditableWrapper = (props) => {
 
 	const handleKeyDown = useCallback((ev) => {
 		const {keyCode, repeat, target} = ev;
-		const {selectedItem} = mutableRef.current;
+		const {selectedItem, selectedItemLabel} = mutableRef.current;
 		const targetItemNode = findItemNode(target);
 
 		if (is('enter', keyCode) && target.getAttribute('role') !== 'button') {
@@ -351,6 +408,13 @@ const EditableWrapper = (props) => {
 					const orders = finalizeOrders();
 					forwardCustom('onComplete', () => ({orders}))({}, editable);
 					reset();
+
+					setTimeout(() => {
+						announceRef.current.announce(
+							selectedItemLabel + $L('move complete'),
+							true
+						);
+					}, completeAnnounceDelay);
 				} else if (targetItemNode) {
 					startEditing(targetItemNode);
 				}
@@ -453,6 +517,7 @@ const EditableWrapper = (props) => {
 			ref={wrapperRef}
 		>
 			{children}
+			<Announce key="editable-wrapper-announce" ref={announceRef} />
 		</div>
 	);
 };